apps/backend/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=67"]

[project]
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "orcheo>=0.11.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0"
]
description = "Deployment wrapper around the Orcheo FastAPI application"
name = "orcheo-backend"
readme = "README.md"
requires-python = ">=3.12"
version = "0.10.0"

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "ruff",
  "mypy"
]

[project.scripts]
orcheo-backend = "orcheo_backend:create_app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
orcheo = {workspace = true}
